Library_management_system/book_reviews/review_api/serializers.py
from rest_framework.serializers import ModelSerializer
from book_reviews.models import Reviews

class ReviewsSerializer(ModelSerializer):
    class Meta:
        model = Reviews
        fields = '__all__'