requirements/base.txt
django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
django-filter==22.1
djangorestframework-simplejwt==5.2.2